Move Mod Manager to prepare class splits
